Moved SortThreads to individual files, added State in MainWindow
